final before deploy

+++ routes/ideas.js
@@ -47,30 +47,53 @@ router.post('/', async (req, res) => {
 // Update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id)
+    console.log(idea.username, req.body.username)
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      {new: true}
-    )
-    res.send({success: true, data: updatedIdea})
+        {new: true}
+      )
+      return res.json({success: true, data: updatedIdea})
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to update this resource',
+    })
   } catch (error) {
     console.log(error)
-    res.status(500).json({success: false, error: 'Something went wrong.  Try again later.'})
+    res.status(500).json({success: false, error: 'Something went wrong'})
   }
 })
 
 // Delete idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id)
-    res.send({success: true, data: {}})
+    const idea = await Idea.findById(req.params.id)
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id)
+      return res.json({success: true, data: {}})
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorized to delete this resource',
+    })
   } catch (error) {
-    res.status(500).json({success: false, error: 'Something went wrong.  Try again later.'})
+    console.log(error)
+    res.status(500).json({success: false, error: 'Something went wrong'})
   }
 })
